added libraries module and api-gateway. libraries been tested and works, api-gateway has not.

+++ booking/src/test/kotlin/com/library/booking/BookingRestApiTest.kt
@@ -123,7 +123,7 @@ class BookingRestApiTest {
             .get("/$dtoId")
             .then()
             .statusCode(200)
-            .body("data.bookId", equalTo(dto.bookId!!.toInt()))
+            .body("data.bookId", equalTo(dto.bookId!!))
     }
